+++ src/main/java/domain/BlackJackGame.java
@@ -1,4 +1,4 @@
-package demain;
+package domain;
 
 import java.util.*;
