+++ src/main/java/domain/Cards.java
@@ -1,4 +1,4 @@
-package demain;
+package domain;
 
 import java.util.List;
