+++ src/main/java/domain/DealerGameResult.java
@@ -1,4 +1,4 @@
-package demain;
+package domain;
 
 import java.util.ArrayList;
 import java.util.List;
