+++ src/main/java/domain/Participant.java
@@ -1,4 +1,4 @@
-package demain;
+package domain;
 
 public class Participant extends Player{
     private static final int STANDARD_NUMBER = 21;
