+++ src/main/java/domain/ParticipantGameResult.java
@@ -1,4 +1,4 @@
-package demain;
+package domain;
 
 import java.util.Map;
 import java.util.Set;
