+++ src/main/java/domain/Participants.java
@@ -1,4 +1,4 @@
-package demain;
+package domain;
 
 import java.util.ArrayList;
 import java.util.List;
