+++ src/test/java/domain/CardNumberTest.java
@@ -1,6 +1,5 @@
 package domain;
 
-import demain.CardNumber;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
