+++ src/test/java/domain/CardTest.java
@@ -1,6 +1,5 @@
 package domain;
 
-import demain.Card;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
