+++ src/test/java/domain/CardsTest.java
@@ -1,7 +1,5 @@
 package domain;
 
-import demain.Card;
-import demain.Cards;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
