+++ src/test/java/domain/DealerGameResultTest.java
@@ -1,7 +1,5 @@
 package domain;
 
-import demain.DealerGameResult;
-import demain.GameResultType;
 import org.junit.jupiter.api.Test;
 
 import java.util.*;
