+++ src/test/java/domain/ParticipantTest.java
@@ -1,9 +1,5 @@
 package domain;
 
-import demain.Card;
-import demain.Cards;
-import demain.GameResultType;
-import demain.Participant;
 import org.junit.jupiter.api.Test;
 
 import java.util.*;
