+++ src/test/java/domain/RandomCardCreatorTest.java
@@ -1,7 +1,5 @@
 package domain;
 
-import demain.Card;
-import demain.RandomCardCreator;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
